+++ src/main/java/com/github/jlangch/venice/impl/VeniceInterpreter.java
@@ -1025,9 +1025,15 @@ public class VeniceInterpreter implements Serializable  {
 			
 			expandedMacros++; 
 
+			final long nanosRun = meterRegistry.enabled ? System.nanoTime() : 0L;
+			
 			ast_ = macro.apply(((VncList)ast_).rest());
+			
+			if (meterRegistry.enabled) {
+				meterRegistry.record(macro.getQualifiedName() + "[m]", System.nanoTime() - nanosRun);
+			}
 		}
-	
+	 
 		if (expandedMacros > 0) {
 			macroExpandCount.addAndGet(expandedMacros);
 			if (meterRegistry.enabled) {