+++ ChangeLog.md
@@ -36,6 +36,10 @@ and simplify code. See the Venice *JMH* benchmark *ReflectionBenchmark.java*
 
 - macros to use the current namespace when expanding as opposed to functions
   that use the namespace they have defined in when executing the body.
+- the REPL to leave char escaping to Venice Reader. Now (print "abc\ndef")  
+  works fine. No double escape (print "abc\\ndef") needed anymore.
+- JLine3 from throwing an IllegalStateException in the REPL while an escaped 
+  char is typed manually.